requirements.txt
boto3>=1.28.0
PyGithub>=2.1.0
PyYAML>=6.0
